# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface ApiError {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface SessionUser {
  id: number;
  name: string;
  username: string;
  email?: string | null;
  role: string;
  localId?: number | null;
  localNombre?: string | null;
}

export interface Sociedad {
  id: number;
  nombre: string;
  cedulaJuridica?: string | null;
  correo?: string | null;
  telefono?: string | null;
  direccion?: string | null;
  activo: boolean;
  localesCount: number;
}

export interface SociedadInput {
  nombre: string;
  cedulaJuridica?: string | null;
  correo?: string | null;
  telefono?: string | null;
  direccion?: string | null;
  activo?: boolean;
}

export interface Marca {
  id: number;
  nombre: string;
  slug?: string | null;
  activo: boolean;
  localesCount: number;
}

export interface MarcaInput {
  nombre: string;
  slug?: string | null;
  activo?: boolean;
}

export interface Local {
  id: number;
  marcaId?: number | null;
  sociedadId?: number | null;
  marcaNombre?: string | null;
  sociedadNombre?: string | null;
  nombre: string;
  codigo?: string | null;
  correo?: string | null;
  telefono?: string | null;
  direccion?: string | null;
  activo: boolean;
}

export interface LocalInput {
  marcaId?: number | null;
  sociedadId?: number | null;
  nombre: string;
  codigo?: string | null;
  correo?: string | null;
  telefono?: string | null;
  direccion?: string | null;
  activo?: boolean;
}

export interface Categoria {
  id: number;
  nombre: string;
  descripcion?: string | null;
  activo: boolean;
  productosCount: number;
}

export interface CategoriaInput {
  nombre: string;
  descripcion?: string | null;
  activo?: boolean;
}

export interface Producto {
  id: number;
  categoriaId?: number | null;
  categoriaNombre?: string | null;
  nombre: string;
  sku?: string | null;
  descripcion?: string | null;
  precio: number;
  costo: number;
  stockMinimo: number;
  activo: boolean;
}

export interface ProductoInput {
  categoriaId?: number | null;
  nombre: string;
  sku?: string | null;
  descripcion?: string | null;
  precio: number;
  costo?: number;
  stockMinimo?: number;
  activo?: boolean;
}

export interface StockItem {
  productoId: number;
  productoNombre: string;
  sku?: string | null;
  categoriaNombre?: string | null;
  localId: number;
  localNombre: string;
  cantidad: number;
  stockMinimo: number;
  /** ok | bajo | agotado */
  estado: string;
}

export interface Movimiento {
  id: number;
  productoId: number;
  productoNombre: string;
  localId: number;
  localNombre: string;
  /** entrada | salida | ajuste | venta */
  tipo: string;
  cantidad: number;
  nota?: string | null;
  usuarioNombre?: string | null;
  createdAt: string;
}

export interface MovimientoInput {
  productoId: number;
  localId: number;
  tipo: string;
  cantidad: number;
  nota?: string | null;
}

export interface PedidoDetalle {
  id: number;
  productoId: number;
  productoNombre: string;
  cantidad: number;
  precioUnitario: number;
  subtotal: number;
}

export interface PedidoDetalleInput {
  productoId: number;
  cantidad: number;
  precioUnitario: number;
}

export interface Pedido {
  id: number;
  localId?: number | null;
  localNombre?: string | null;
  clienteNombre?: string | null;
  clienteTelefono?: string | null;
  clienteEmail?: string | null;
  observaciones?: string | null;
  subtotal: number;
  impuesto: number;
  total: number;
  estado: string;
  createdAt: string;
  detalles: PedidoDetalle[];
}

export interface PedidoInput {
  localId?: number | null;
  clienteNombre?: string | null;
  clienteTelefono?: string | null;
  clienteEmail?: string | null;
  observaciones?: string | null;
  impuestoPct?: number;
  detalles: PedidoDetalleInput[];
}

export interface FacturaDetalle {
  id: number;
  productoId?: number | null;
  descripcion: string;
  cantidad: number;
  precioUnitario: number;
  subtotal: number;
}

export interface Factura {
  id: number;
  pedidoId?: number | null;
  localId?: number | null;
  localNombre?: string | null;
  numeroFactura: string;
  clienteNombre?: string | null;
  clienteTelefono?: string | null;
  clienteEmail?: string | null;
  fecha: string;
  subtotal: number;
  impuesto: number;
  total: number;
  estado: string;
  observaciones?: string | null;
  detalles: FacturaDetalle[];
}

export interface Usuario {
  id: number;
  name: string;
  username: string;
  email?: string | null;
  role: string;
  localId?: number | null;
  localNombre?: string | null;
}

export interface UsuarioInput {
  name: string;
  username: string;
  email?: string | null;
  password?: string | null;
  role: string;
  localId?: number | null;
}

export interface DashboardResumen {
  ventasTotales: number;
  pedidosCount: number;
  facturasCount: number;
  productosCount: number;
  ticketPromedio: number;
  stockBajoCount: number;
}

export interface VentaPorDia {
  fecha: string;
  total: number;
  pedidos: number;
}

export interface TopProducto {
  productoId: number;
  productoNombre: string;
  cantidadVendida: number;
  ingresos: number;
}

export interface VentaPorLocal {
  localId: number;
  localNombre: string;
  marcaNombre?: string | null;
  total: number;
  pedidos: number;
}

export interface StockBajoItem {
  productoId: number;
  productoNombre: string;
  localId: number;
  localNombre: string;
  cantidad: number;
  stockMinimo: number;
}

export interface ActividadItem {
  tipo: string;
  id: number;
  titulo: string;
  subtitulo?: string | null;
  monto?: number | null;
  createdAt: string;
}

export type GetDashboardResumenParams = {
  dias?: number;
};

export type GetVentasPorDiaParams = {
  dias?: number;
};

export type GetTopProductosParams = {
  dias?: number;
  limite?: number;
};

export type GetVentasPorLocalParams = {
  dias?: number;
};

export type GetStockBajoParams = {
  limite?: number;
};

export type ListProductosParams = {
  search?: string;
  categoriaId?: number;
};

export type ListStockParams = {
  localId?: number;
  search?: string;
};

export type ListMovimientosParams = {
  localId?: number;
  limite?: number;
};

export type ListPedidosParams = {
  estado?: string;
  localId?: number;
};

export type ListFacturasParams = {
  localId?: number;
  estado?: string;
};