# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHF Inventarios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Usuario autenticado
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ------------------ DASHBOARD ------------------
  /dashboard/resumen:
    get:
      operationId: getDashboardResumen
      tags: [dashboard]
      summary: Métricas resumen
      parameters:
        - in: query
          name: dias
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResumen"
  /dashboard/ventas-por-dia:
    get:
      operationId: getVentasPorDia
      tags: [dashboard]
      summary: Serie de ventas por día
      parameters:
        - in: query
          name: dias
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VentaPorDia" }
  /dashboard/top-productos:
    get:
      operationId: getTopProductos
      tags: [dashboard]
      summary: Productos más vendidos
      parameters:
        - in: query
          name: dias
          schema: { type: integer, default: 30 }
        - in: query
          name: limite
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TopProducto" }
  /dashboard/ventas-por-local:
    get:
      operationId: getVentasPorLocal
      tags: [dashboard]
      summary: Ventas agrupadas por local
      parameters:
        - in: query
          name: dias
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VentaPorLocal" }
  /dashboard/stock-bajo:
    get:
      operationId: getStockBajo
      tags: [dashboard]
      summary: Productos con stock bajo
      parameters:
        - in: query
          name: limite
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StockBajoItem" }
...
  /categorias:
    get:
      operationId: listCategorias
      tags: [categorias]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Categoria" }
    post:
      operationId: createCategoria
      tags: [categorias]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoriaInput" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Categoria" }
  /categorias/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    put:
      operationId: updateCategoria
      tags: [categorias]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoriaInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Categoria" }
    delete:
      operationId: deleteCategoria
      tags: [categorias]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  # ------------------ PRODUCTOS ------------------
...
  /facturas:
    get:
      operationId: listFacturas
      tags: [facturas]
      parameters:
        - in: query
          name: localId
          schema: { type: integer }
        - in: query
          name: estado
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Factura" }
  /facturas/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getFactura
      tags: [facturas]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Factura" }

  # ------------------ USUARIOS ------------------
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    ApiError:
      type: object
      properties:
        message: { type: string }
      required: [message]
    LoginBody:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    SessionUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        username: { type: string }
        email: { type: string, nullable: true }
        role: { type: string }
        localId: { type: integer, nullable: true }
        localNombre: { type: string, nullable: true }
      required: [id, name, username, role]
...
    Categoria:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        descripcion: { type: string, nullable: true }
        activo: { type: boolean }
        productosCount: { type: integer }
      required: [id, nombre, activo, productosCount]
    CategoriaInput:
      type: object
      properties:
        nombre: { type: string }
        descripcion: { type: string, nullable: true }
        activo: { type: boolean }
      required: [nombre]
...
    FacturaDetalle:
      type: object
      properties:
        id: { type: integer }
        productoId: { type: integer, nullable: true }
        descripcion: { type: string }
        cantidad: { type: integer }
        precioUnitario: { type: number }
        subtotal: { type: number }
      required: [id, descripcion, cantidad, precioUnitario, subtotal]
    Factura:
      type: object
      properties:
        id: { type: integer }
        pedidoId: { type: integer, nullable: true }
        localId: { type: integer, nullable: true }
        localNombre: { type: string, nullable: true }
        numeroFactura: { type: string }
        clienteNombre: { type: string, nullable: true }
        clienteTelefono: { type: string, nullable: true }
        clienteEmail: { type: string, nullable: true }
        fecha: { type: string }
        subtotal: { type: number }
        impuesto: { type: number }
        total: { type: number }
        estado: { type: string }
        observaciones: { type: string, nullable: true }
        detalles:
          type: array
          items: { $ref: "#/components/schemas/FacturaDetalle" }
      required: [id, numeroFactura, fecha, subtotal, impuesto, total, estado, detalles]
...
    DashboardResumen:
      type: object
      properties:
        ventasTotales: { type: number }
        pedidosCount: { type: integer }
        facturasCount: { type: integer }
        productosCount: { type: integer }
        ticketPromedio: { type: number }
        stockBajoCount: { type: integer }
      required: [ventasTotales, pedidosCount, facturasCount, productosCount, ticketPromedio, stockBajoCount]
    VentaPorDia:
      type: object
      properties:
        fecha: { type: string }
        total: { type: number }
        pedidos: { type: integer }
      required: [fecha, total, pedidos]
    TopProducto:
      type: object
      properties:
        productoId: { type: integer }
        productoNombre: { type: string }
        cantidadVendida: { type: integer }
        ingresos: { type: number }
      required: [productoId, productoNombre, cantidadVendida, ingresos]
    VentaPorLocal:
      type: object
      properties:
        localId: { type: integer }
        localNombre: { type: string }
        marcaNombre: { type: string, nullable: true }
        total: { type: number }
        pedidos: { type: integer }
      required: [localId, localNombre, total, pedidos]
    StockBajoItem:
      type: object
      properties:
        productoId: { type: integer }
        productoNombre: { type: string }
        localId: { type: integer }
        localNombre: { type: string }
        cantidad: { type: integer }
        stockMinimo: { type: integer }
      required: [productoId, productoNombre, localId, localNombre, cantidad, stockMinimo]
...