# lib/api-zod/src/generated/types/categoriaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoriaInput {
  nombre: string;
  descripcion?: string | null;
  activo?: boolean;
}