# lib/api-zod/src/generated/types/facturaDetalle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface FacturaDetalle {
  id: number;
  productoId?: number | null;
  descripcion: string;
  cantidad: number;
  precioUnitario: number;
  subtotal: number;
}