# lib/api-zod/src/generated/types/ventaPorDia.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GHF Inventarios API
 * OpenAPI spec version: 0.1.0
 */

export interface VentaPorDia {
  fecha: string;
  total: number;
  pedidos: number;
}